requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
idna==2.9
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
